fix issues with index creation if already existing
Neo4j 4.1.3 introduces a handy "IF EXISTS" optional modifier on the CREATE
INDEX clause. However, for anyone running the still-supported version of
4.0, this isn't available :-(

This adds a try/catch block and checks if the failure from the index
creation statement was because of an existing index. If not, it
raises the exception. Otherwise it silently ignores it.